rides overview and maps - working flow

+++ Domain/Models/Location.cs
@@ -6,6 +6,6 @@ public class Location
     public string City { get; set; }
     public string StreetName { get; set; }
     public string ZipCode { get; set; }
-    public long CoordinatesX { get; set; }
-    public long CoordinatesY { get; set; }
+    public double CoordinatesX { get; set; }
+    public double CoordinatesY { get; set; }
 }
